+++ Runtime/Components/GlyphHybridOutput.cs
@@ -3,15 +3,16 @@ using TMPro;
 using UnityEngine;
 using UnityEngine.UI;
 
-namespace LMirman.RewiredGlyphs
+namespace LMirman.RewiredGlyphs.Components
 {
 	/// <summary>
-	/// A <see cref="InputGlyphDisplay"/> element that will set it's <see cref="textMesh"/> and <see cref="image"/> components to the <see cref="InputGlyphDisplay.actionName"/>'s <see cref="InputGlyph.Description"/> and <see cref="InputGlyph.GetSprite(AxisRange)"/> values respectively.
+	/// A <see cref="GlyphDisplay"/> element that will set it's <see cref="textMesh"/> and <see cref="image"/> components to the <see cref="GlyphDisplay.actionName"/>'s <see cref="Glyph.Description"/> and <see cref="Glyph.GetSprite(AxisRange)"/> values respectively.
 	/// </summary>
 	/// <remarks>
 	/// Using the <see cref="alwaysRenderBoth"/> as true can show both sprite and description elements where using it as false can utilize the text component as a fallback if the InputGlyph does not have a sprite defined.
 	/// </remarks>
-	public class InputGlyphHybrid : InputGlyphDisplay
+	[AddComponentMenu("Rewired Glyphs/Image and Text Hybrid Output")]
+	public class GlyphHybridOutput : GlyphDisplay
 	{
 		[SerializeField]
 		private bool alwaysRenderBoth;
@@ -25,7 +26,7 @@ namespace LMirman.RewiredGlyphs
 			image = GetComponentInChildren<Image>();
 		}
 
-		public override void SetGlyph(InputGlyph glyph, AxisRange axisRange)
+		public override void SetGlyph(Glyph glyph, AxisRange axisRange)
 		{
 			Sprite sprite = glyph.GetSprite(axisRange);
 			image.enabled = sprite != null;