+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Added
 - Input glyphs can now be retrieved on a per player basis.
+- `GlyphDisplay` can now use action ids instead of names.
 
 ### Changes
 - `GenericEntry` on `InputGlyphCollection` has been renamed to `TemplateEntry`
@@ -16,8 +17,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - Enums that were previously nested within `InputGlyphs` have been moved to outer scope
 - Everything in the package is now within the `LMirman.RewiredGlyphs` namespace
 - InputGlyphs now uses an integer parameter for actions instead of string
+- Most class names containing `InputGlyph` have been simplified to just `Glyph`
+  - `InputGlyphs` will not be renamed.
 
 ## [0.1.0] - 2023-06-20
 
 ### Added
-- Package created.
+- Package created.~~~~